// back/src/shared/enums/error.enum.ts
export enum EGenericError {
  TICKET_IS_CLOSED = 'generic.TICKET_IS_CLOSED',
  PROJECT_ID_NOT_FOUND = 'generic.PROJECT_ID_NOT_FOUND',
  TICKET_NOT_SIGN = 'generic.TICKET_NOT_SIGN',
  TICKET_HAS_ANALYST = 'generic.TICKET_HAS_ANALYST',
  TICKET_ALREADY_HAS_AN_OWNER = 'generic.TICKET_ALREADY_HAS_AN_OWNER',
  FORBIDDEN = 'generic.FORBIDDEN',
  UNAUTHORIZED = 'generic.UNAUTHORIZED',
  NOT_ALLOWED = 'generic.NOT_ALLOWED',
  INTERNAL_SERVER_ERROR = 'generic.INTERNAL_SERVER_ERROR',
  BRANCH_NOT_FOUND = 'generic.BRANCH_NOT_FOUND',
  DOCKER_OFFLINE = 'generic.DOCKER_OFFLINE',
  ROUTER_NAME_EXISTS = 'generic.ROUTER_NAME_EXISTS',
  ROUTER_NAME_NOT_FOUND = 'generic.ROUTER_NAME_NOT_FOUND',
  SERVICE_NAME_EXISTS = 'generic.SERVICE_NAME_EXISTS',
  SERVICE_NAME_NOT_FOUND = 'generic.SERVICE_NAME_NOT_FOUND',
  CURRENT_PASSWORD_INVALID = 'generic.CURRENT_PASSWORD_INVALID',
  GITHUB_USER_NOT_FOUND = 'generic.GITHUB_USER_NOT_FOUND',
  NO_ACCOUNT_LINKED = 'generic.NO_ACCOUNT_LINKED',
  ID_EXISTS = 'generic.ID_EXISTS',
}